25_10_2023

+++ c_plus_plus/data_types/type_sizes.cpp
@@ -2,6 +2,6 @@
 #include <limits>
 
 int main() {
-    std::cout << "minimum value: " << std::numeric_limits<short int>::min() << "\n"
-    << "maximum value: " << std::numeric_limits<short int>::max() << "\n";
+    std::cout << "minimum value: " << std::numeric_limits<int>::min() << "\n"
+    << "maximum value: " << std::numeric_limits<int>::max() << "\n";
 }
